On branch main Your branch is up to date with 'origin/main'.
Changes to be committed:
	modified:   app/Http/Controllers/admin/BlogController.php
	modified:   app/Http/Controllers/admin/KomentarBlogController.php
	modified:   app/Http/Controllers/admin/KomentarProjectController.php
	modified:   app/Http/Controllers/admin/ProjectController.php
	modified:   app/Http/Controllers/admin/SettingController.php
	modified:   app/Http/Controllers/frontpage/BlogController.php
	modified:   app/Http/Controllers/frontpage/ProjectController.php
	modified:   app/Models/admin/Project.php
	deleted:    public/administrator/assets/media/blog/image__2023-12-10-00-32-03_26574a49307156.jpg
	deleted:    public/administrator/assets/media/blog/image_pengembala-oray_2023-12-10-00-54-16_26574a9c8d6d21.jpg
	deleted:    public/administrator/assets/media/blog/image_upin-ipin_2023-12-10-01-10-38_26574ad9e00d98.jpg
	deleted:    public/administrator/assets/media/blog/isi_pengembala-oray-17021444560.png
	deleted:    public/administrator/assets/media/project/image_portalow_2023-12-05-05-54-11_2656e5893d479e.jpg
	deleted:    public/administrator/assets/media/project/image_portalow_2023-12-05-05-54-14_2656e58961f036.jpg
	deleted:    public/administrator/assets/media/project/image_portalow_2023-12-05-05-54-16_2656e589834c28.jpg
	deleted:    public/administrator/assets/media/project/image_portalow_2023-12-05-10-48-11_2656e9d7b697de.jpg
	modified:   resources/views/administrator/blog/detail.blade.php
	modified:   resources/views/administrator/layouts/nav.blade.php
	modified:   resources/views/administrator/project/arsip.blade.php
	modified:   resources/views/administrator/project/detail.blade.php
	modified:   resources/views/administrator/settings/admin/index.blade.php
	new file:   resources/views/administrator/settings/administrator.blade.php
	new file:   resources/views/administrator/settings/frontpage.blade.php
	modified:   resources/views/administrator/settings/frontpage/footer.blade.php
	modified:   resources/views/administrator/settings/frontpage/general.blade.php
	new file:   resources/views/administrator/settings/frontpage/homepage.blade.php
	modified:   resources/views/administrator/settings/main.blade.php
	modified:   resources/views/frontpage/home/index.blade.php
	modified:   routes/admin.php

+++ app/Http/Controllers/admin/KomentarProjectController.php
@@ -58,19 +58,20 @@ class KomentarProjectController extends Controller
 
         // Find the data based on the provided ID or throw a 404 exception.
         $data = KomentarProject::with('reply')->findOrFail($id);
+        if (!$data) {
+            abort(404);
+        }
 
         // Store the data to be logged before deletion
         $deletedData = $data->toArray();
 
-        $detail = [];
-        foreach ($data->reply as $row) {
-            $detail[] = $row;
-            $row->delete();
+        if(!$data->reply->isEmpty()){
+            $data->reply->each->delete();
         }
 
         $dataJson = [
             'data' => $deletedData,
-            'reply' => $detail
+            'reply' => $data->reply
         ];
         // Delete the data.
         $data->delete();
@@ -98,9 +99,10 @@ class KomentarProjectController extends Controller
         return view('administrator.komentar_project.detail',compact('data'));
     }
 
-    public function getDataDetail(Request $request){
+    public function getDataDetail(Request $request, $id){
         $data = KomentarProjectReply::query()
-                                ->with('project');
+                                ->with('project')
+                                ->where('komentar_id', $id);
 
         $data = $data->get();
 
@@ -129,7 +131,9 @@ class KomentarProjectController extends Controller
 
         // Find the data based on the provided ID or throw a 404 exception.
         $data = KomentarProjectReply::findOrFail($id);
-
+        if (!$data) {
+            abort(404);
+        }
         // Store the data to be logged before deletion
         $deletedData = $data->toArray();
